+++ samples/standalone/src/main/kotlin/org/jetbrains/jewel/samples/standalone/viewmodel/ComponentsViewModel.kt
@@ -9,7 +9,7 @@ import androidx.compose.runtime.snapshots.SnapshotStateList
 import org.jetbrains.jewel.samples.showcase.components.Banners
 import org.jetbrains.jewel.samples.standalone.StandaloneSampleIcons
 import org.jetbrains.jewel.samples.showcase.components.Borders
-import org.jetbrains.jewel.samples.standalone.view.component.Buttons
+import org.jetbrains.jewel.samples.showcase.components.Buttons
 import org.jetbrains.jewel.samples.standalone.view.component.Checkboxes
 import org.jetbrains.jewel.samples.standalone.view.component.ChipsAndTrees
 import org.jetbrains.jewel.samples.standalone.view.component.Dropdowns
